# docs/swagger.yaml
definitions:
  api.BroadcastRequest:
    properties:
      message:
        type: string
    type: object
...
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  api.MessageResponse:
    properties:
      message:
        type: string
    type: object
  api.SendRconCommandRequest:
    properties:
      content:
        type: string
      uuid:
        type: string
    type: object
  api.ServerInfo:
    properties:
      name:
        type: string
      version:
        type: string
    type: object
...
  api.ShutdownRequest:
    properties:
      message:
        type: string
      seconds:
        type: integer
    type: object
  api.SuccessResponse:
    properties:
      success:
        type: boolean
    type: object
...
  database.RconCommand:
    properties:
      command:
        type: string
      placeholder:
        type: string
      remark:
        type: string
    type: object
  database.RconCommandList:
    properties:
      command:
        type: string
      placeholder:
        type: string
      remark:
        type: string
      uuid:
        type: string
    type: object
...
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /api/rcon:
    get:
      consumes:
      - application/json
      description: List Rcon Commands
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.RconCommandList'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List Rcon Commands
      tags:
      - Rcon
    post:
      consumes:
      - application/json
      description: Add Rcon Command
      parameters:
      - description: Rcon Command
        in: body
        name: command
        required: true
        schema:
          $ref: '#/definitions/database.RconCommand'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Add Rcon Command
      tags:
      - Rcon
  /api/rcon/{uuid}:
    delete:
      consumes:
      - application/json
      description: Remove Rcon Command
      parameters:
      - description: UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Remove Rcon Command
      tags:
      - Rcon
    put:
      consumes:
      - application/json
      description: Put Rcon Command
      parameters:
      - description: UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: Rcon Command
        in: body
        name: command
        required: true
        schema:
          $ref: '#/definitions/database.RconCommand'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Put Rcon Command
      tags:
      - Rcon
  /api/rcon/import:
    post:
      consumes:
      - multipart/form-data
      description: Import Rcon Commands from a TXT file
      parameters:
      - description: Upload txt file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Import Rcon Commands
      tags:
      - Rcon
  /api/rcon/send:
    post:
      consumes:
      - application/json
      description: Send Rcon Command
      parameters:
      - description: Rcon Command
        in: body
        name: command
        required: true
        schema:
          $ref: '#/definitions/api.SendRconCommandRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Send Rcon Command
      tags:
      - Rcon
  /api/server:
    get:
      consumes:
      - application/json
      description: Get Server Info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ServerInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get Server Info
      tags:
      - Server
  /api/server/broadcast:
    post:
      consumes:
      - application/json
      description: Publish Broadcast
      parameters:
      - description: Broadcast
        in: body
        name: broadcast
        required: true
        schema:
          $ref: '#/definitions/api.BroadcastRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Publish Broadcast
      tags:
      - Server
...
  /api/server/shutdown:
    post:
      consumes:
      - application/json
      description: Shutdown Server
      parameters:
      - description: Shutdown
        in: body
        name: shutdown
        required: true
        schema:
          $ref: '#/definitions/api.ShutdownRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Shutdown Server
      tags:
      - Server
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"